% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-summary.R
\name{add_class_icon}
\alias{add_class_icon}
\title{Convert class grid column to icon}
\usage{
add_class_icon(grid, column = "class")
}
\arguments{
\item{grid}{grid}

\item{column}{column}
}
\value{
datagrid
}
\description{
Convert class grid column to icon
}
\examples{
mtcars |>
  overview_vars() |>
  toastui::datagrid() |>
  add_class_icon()
}
