% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-summary.R
\name{add_sparkline}
\alias{add_sparkline}
\title{Add sparkline to datagrid}
\usage{
add_sparkline(
  grid,
  column = "vals",
  color.main = "#2a8484",
  color.sec = "#84EF84"
)
}
\arguments{
\item{grid}{grid}

\item{column}{clumn to transform}
}
\value{
datagrid
}
\description{
Add sparkline to datagrid
}
\examples{
grid <- mtcars |>
  default_parsing() |>
  overview_vars() |>
  toastui::datagrid() |>
  add_sparkline()
grid
}
