% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plots.R
\name{allign_axes}
\alias{allign_axes}
\title{Alligns axes between plots}
\usage{
allign_axes(...)
}
\arguments{
\item{...}{ggplot2 objects or list of ggplot2 objects}
}
\value{
list of ggplot2 objects
}
\description{
Alligns axes between plots
}
