% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{append_list}
\alias{append_list}
\title{Append list with named index}
\usage{
append_list(data, list, index)
}
\arguments{
\item{data}{data to add to list}

\item{list}{list}

\item{index}{index name}
}
\value{
list
}
\description{
Append list with named index
}
\examples{
ls_d <- list(test = c(1:20))
ls_d <- list()
data.frame(letters[1:20], 1:20) |> append_list(ls_d, "letters")
letters[1:20] |> append_list(ls_d, "letters")
}
