% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{available_plots}
\alias{available_plots}
\title{Implemented functions}
\usage{
available_plots()
}
\value{
list
}
\description{
Library of supported functions. The list name and "descr" element should be
unique for each element on list.
\itemize{
\item fun: the plotting function
\item fun.args: default parameters for the plotting function
\item descr: Plot description
\item note: Short note/description of the function for displaying in ui and docs
\item primary.type: Primary variable data type (see \link{data_type})
\item base: holds a list of parameters for plot input fields generation
Secondary and tertiary variable input fields are mandatory.
}
}
\examples{
available_plots() |> str()
}
