% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-summary.R
\name{class_icons}
\alias{class_icons}
\title{Get data class icons}
\usage{
class_icons(x)
}
\arguments{
\item{x}{character vector of data classes}
}
\value{
list
}
\description{
Get data class icons
}
\examples{
"numeric" |> class_icons()|> str()
mtcars |> sapply(class) |> class_icons() |> str()
}
