% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-variables-ext.R
\name{clean_date}
\alias{clean_date}
\title{Attempts at applying uniform date format}
\usage{
clean_date(data)
}
\arguments{
\item{data}{character string vector of possible dates}
}
\value{
character string
}
\description{
Attempts at applying uniform date format
}
