% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-variables-ext.R
\name{clean_sep}
\alias{clean_sep}
\title{gsub wrapper for piping with default values for separator substituting}
\usage{
clean_sep(data, old.sep = "[-.,/]", new.sep = "-")
}
\arguments{
\item{data}{character vector}

\item{old.sep}{old separator}

\item{new.sep}{new separator}
}
\value{
character vector
}
\description{
gsub wrapper for piping with default values for separator substituting
}
