% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missings-module.R
\name{compare_missings}
\alias{compare_missings}
\title{Pairwise comparison of missings across covariables}
\usage{
compare_missings(data, by_var, max_level = 20)
}
\arguments{
\item{data}{data frame}

\item{by_var}{variable to stratify by missingness}
}
\value{
gtsummary list object
}
\description{
Pairwise comparison of missings across covariables
}
