% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{corr_pairs_validate}
\alias{corr_pairs_validate}
\title{Correlation pairs validation}
\usage{
corr_pairs_validate(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
data.frame
}
\description{
Correlation pairs validation
}
\examples{
# correlation_pairs(mtcars) |> corr_pairs_validate()
}
