% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations-module.R
\name{correlation_pairs}
\alias{correlation_pairs}
\title{Determine significant correlations in the data set}
\usage{
correlation_pairs(data, threshold = 0.8)
}
\arguments{
\item{data}{data.frame}

\item{threshold}{correlation threshold}
}
\value{
data.frame
}
\description{
Determine significant correlations in the data set
}
\examples{
correlation_pairs(mtcars)
correlation_pairs(mtcars,.9)
correlation_pairs(mtcars[c(1:4),])
}
