% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_table.R
\name{create_baseline}
\alias{create_baseline}
\title{Create a baseline table}
\usage{
create_baseline(
  data,
  ...,
  by.var,
  add.p = FALSE,
  add.overall = FALSE,
  theme = c("jama", "lancet", "nejm", "qjecon")
)
}
\arguments{
\item{data}{data}

\item{...}{passed as fun.arg to baseline_table()}

\item{add.p}{add comparison/p-value}

\item{add.overall}{add overall column}

\item{strat.var}{grouping/strat variable}
}
\value{
gtsummary table list object
}
\description{
Create a baseline table
}
\examples{
mtcars |> create_baseline(by.var = "gear", add.p = "yes" == "yes")
create_baseline(default_parsing(mtcars), by.var = "am", add.p = FALSE, add.overall = FALSE, theme = "lancet")
}
