% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plot.R
\name{create_log_tics}
\alias{create_log_tics}
\title{Create summetric log ticks}
\usage{
create_log_tics(data)
}
\arguments{
\item{data}{numeric vector}
}
\value{
numeric vector
}
\description{
Create summetric log ticks
}
\examples{
c(sample(seq(.1, 1, .1), 3), sample(1:10, 3)) |> create_log_tics()
}
