% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-summary.R
\name{create_overview_datagrid}
\alias{create_overview_datagrid}
\title{Create a data overview datagrid}
\usage{
create_overview_datagrid(data, ...)
}
\arguments{
\item{data}{data}
}
\value{
datagrid
}
\description{
Create a data overview datagrid
}
\examples{
mtcars |>
  overview_vars() |>
  create_overview_datagrid()
}
