% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations-module.R, R/data_plots.R
\name{data-correlations}
\alias{data-correlations}
\alias{data_correlations_ui}
\alias{data_correlations_server}
\alias{data_visuals_ui}
\alias{data_visuals_server}
\title{Data correlations evaluation module}
\usage{
data_correlations_ui(id, ...)

data_correlations_server(id, data, include.class = NULL, cutoff = 0.7, ...)

data_visuals_ui(id, tab_title = "Plots", ...)

data_visuals_server(id, data, ...)
}
\arguments{
\item{id}{Module id. (Use 'ns("id")')}

\item{...}{ignored}

\item{data}{data}

\item{color.main}{main color}

\item{color.sec}{secondary color}
}
\value{
Shiny ui module

shiny server module

Shiny ui module

shiny server module
}
\description{
Data correlations evaluation module

Data correlations evaluation module
}
