% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-import.R
\name{data_import_demo_app}
\alias{data_import_demo_app}
\title{Test app for the data-import module}
\usage{
data_import_demo_app()
}
\description{
Test app for the data-import module
}
\examples{
\dontrun{
data_import_demo_app()
}
}
