% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{data_description}
\alias{data_description}
\title{Ultra short data dascription}
\usage{
data_description(data)
}
\arguments{
\item{data}{}
}
\value{
character vector
}
\description{
Ultra short data dascription
}
\examples{
data.frame(
  sample(1:8, 20, TRUE),
  sample(c(1:8, NA), 20, TRUE)
) |> data_description()
}
