% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_summary.R
\name{data_summary_gather}
\alias{data_summary_gather}
\title{Data summary for printing visual summary}
\usage{
data_summary_gather(
  data,
  summary.fun = class,
  palette.fun = viridisLite::viridis
)
}
\arguments{
\item{data}{data.frame}

\item{palette.fun}{optionally use specific palette functions. First argument
has to be the length.}

\item{fun}{summary function. Default is "class"}
}
\value{
data.frame
}
\description{
Data summary for printing visual summary
}
\examples{
mtcars |> data_summary_gather()
}
