% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{data_type_filter}
\alias{data_type_filter}
\title{Filter function to filter data set by variable type}
\usage{
data_type_filter(data, type)
}
\arguments{
\item{data}{data frame}

\item{type}{vector of data types (recognised: data_types)}
}
\value{
data.frame
}
\description{
Filter function to filter data set by variable type
}
\examples{
default_parsing(mtcars) |> data_type_filter(type=c("categorical","continuous")) |> attributes()
\dontrun{
default_parsing(mtcars) |> data_type_filter(type=c("test","categorical","continuous"))
}
}
