% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_string.R
\name{detect_delimiter}
\alias{detect_delimiter}
\title{Detect delimiters in string based on allowed delimiters}
\usage{
detect_delimiter(data, delimiters = c("_", "-", ";", "\\n", ","))
}
\arguments{
\item{delimiters}{allowed delimiters}

\item{text}{character vector}
}
\value{
character vector
}
\description{
Accepts any repeat of delimiters and includes surrounding whitespace
}
\examples{
sapply(c("Walk - run", "Sel__Re", "what;now"), detect_delimiter)
}
