% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{dim_change_call}
\alias{dim_change_call}
\title{Dimensions validation}
\usage{
dim_change_call(before, after, fun)
}
\arguments{
\item{before}{data before}

\item{after}{data after}

\item{fun}{dimension function. ncol or nrow}
}
\value{
data.frame
}
\description{
Dimensions validation
}
