% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_check.R
\name{.has_internet}
\alias{.has_internet}
\title{Check internet connectivity}
\usage{
.has_internet()
}
\value{
Logical; TRUE if an internet connection is available.
}
\description{
Check internet connectivity
}
