% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{expression_string}
\alias{expression_string}
\title{Deparses expression as string, substitutes native pipe and adds assign}
\usage{
expression_string(data, assign.str = "")
}
\arguments{
\item{data}{expression}
}
\value{
string
}
\description{
Deparses expression as string, substitutes native pipe and adds assign
}
\examples{
list(
  as.symbol(paste0("mtcars$", "mpg")),
  rlang::call2(.fn = "select", !!!list(c("cyl", "disp")), .ns = "dplyr"),
  rlang::call2(.fn = "default_parsing", .ns = "FreesearchR")
) |>
  merge_expression() |>
  expression_string()
}
