% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{FreesearchR_palette}
\alias{FreesearchR_palette}
\title{Use the FreesearchR colors}
\usage{
FreesearchR_palette(n)
}
\arguments{
\item{n}{number of colors}
}
\value{
character vector
}
\description{
Use the FreesearchR colors
}
\examples{
FreesearchR_palette(n=7)
}
