% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_globalenv-ext.R
\name{get_dimensions}
\alias{get_dimensions}
\title{Extension of the helper function from datamods}
\usage{
get_dimensions(objs, filter_df = TRUE)
}
\arguments{
\item{objs}{objs}

\item{filter_df}{flag to only include data frames}
}
\value{
vector of data frames with the package names as attr
}
\description{
Extension of the helper function from datamods
}
