% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_model.R
\name{get_fun_options}
\alias{get_fun_options}
\title{Get the function options based on the selected function description}
\usage{
get_fun_options(data)
}
\arguments{
\item{data}{vector}
}
\value{
list
}
\description{
Get the function options based on the selected function description
}
\examples{
mtcars |>
  default_parsing() |>
  dplyr::pull(mpg) |>
  possible_functions(design = "cross-sectional") |>
  (\(.x){
    .x[[1]]
  })() |>
  get_fun_options()
}
