% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{get_input_params}
\alias{get_input_params}
\title{Get the function parameters based on the selected function description}
\usage{
get_input_params(data)
}
\arguments{
\item{data}{vector}
}
\value{
list
}
\description{
Get the function parameters based on the selected function description
}
\examples{
ls <- mtcars |>
  default_parsing() |>
  dplyr::pull(mpg) |>
  possible_plots() |>
  (\(.x){
    .x[[1]]
  })() |>
  get_input_params()
}
