% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut-variable-ext.R
\name{get_list_elements}
\alias{get_list_elements}
\title{Subset elements from list of lists}
\usage{
get_list_elements(name, element, dict = cut_methods())
}
\arguments{
\item{name}{list name to lookup}

\item{element}{element to get}

\item{dict}{dictionary to use}
}
\value{
named vector
}
\description{
General function to sub-setting details stored in list dictionaries.
}
\examples{
get_list_elements(c("top", "bottom"), "descr")
}
