% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plots.R
\name{get_plot_options}
\alias{get_plot_options}
\title{Get the function options based on the selected function description}
\usage{
get_plot_options(data)
}
\arguments{
\item{data}{vector}
}
\value{
list
}
\description{
Get the function options based on the selected function description
}
\examples{
ls <- mtcars |>
  default_parsing() |>
  dplyr::pull(mpg) |>
  possible_plots() |>
  (\(.x){
    .x[[1]]
  })() |>
  get_plot_options()
}
