% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-summary.R
\name{get_var_icon}
\alias{get_var_icon}
\title{Easily get variable icon based on data type or class}
\usage{
get_var_icon(data, class.type = c("class", "type"))
}
\arguments{
\item{data}{variable or data frame}

\item{class.type}{"type" or "class". Default is "class"}
}
\value{
svg icon
}
\description{
Easily get variable icon based on data type or class
}
\examples{
mtcars[1] |> get_var_icon("class")
default_parsing(mtcars) |> get_var_icon()
}
