% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{gg_theme_export}
\alias{gg_theme_export}
\title{GGplot default theme for plotting export objects}
\usage{
gg_theme_export()
}
\arguments{
\item{data}{ggplot object}
}
\value{
ggplot object
}
\description{
GGplot default theme for plotting export objects
}
