% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{gg_theme_shiny}
\alias{gg_theme_shiny}
\title{GGplot default theme for plotting in Shiny}
\usage{
gg_theme_shiny()
}
\arguments{
\item{data}{ggplot object}
}
\value{
ggplot object
}
\description{
GGplot default theme for plotting in Shiny
}
