% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_euler.R
\name{ggeulerr}
\alias{ggeulerr}
\title{Area proportional venn diagrams}
\usage{
ggeulerr(combinations, show_quantities = TRUE, show_labels = TRUE, ...)
}
\arguments{
\item{combinations}{set relationships as a named numeric vector, matrix, or
data.frame(See \code{eulerr::euler})}

\item{show_quantities}{whether to show number of intersecting elements}

\item{show_labels}{whether to show set names}

\item{...}{further arguments passed to eulerr::euler}
}
\description{
THis is slightly modified from https://gist.github.com/danlooo/d23d8bcf8856c7dd8e86266097404ded

This functions uses eulerr::euler to plot area proportional venn diagramms
but plots it using ggplot2
}
