% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{if_not_missing}
\alias{if_not_missing}
\title{Return if available}
\usage{
if_not_missing(data, default = NULL)
}
\arguments{
\item{data}{vector}

\item{default}{assigned value for missings}
}
\value{
vector
}
\description{
Return if available
}
\examples{
NULL |> if_not_missing("new")
c(2,"a",NA) |> if_not_missing()
"See" |> if_not_missing()
}
