% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-file-ext.R
\name{import-file-type}
\alias{import-file-type}
\alias{import_delim}
\alias{import_xls}
\alias{import_ods}
\alias{import_dta}
\alias{import_rds}
\title{Wrapper to ease data file import}
\usage{
import_delim(file, skip, encoding, na.strings)

import_xls(file, sheet, skip, na.strings)

import_ods(file, sheet, skip, na.strings)

import_dta(file)

import_rds(file)
}
\arguments{
\item{file}{path to the file}

\item{skip}{number of row to skip}

\item{encoding}{file encoding}

\item{na.strings}{character(s) to interpret as missing values.}

\item{sheet}{for Excel files, sheet to read}
}
\value{
data.frame

data.frame

data.frame

data.frame

data.frame
}
\description{
Wrapper to ease data file import
}
