% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-file-ext.R
\name{import_file_demo_app}
\alias{import_file_demo_app}
\title{Test app for the import_file module}
\usage{
import_file_demo_app()
}
\description{
Test app for the import_file module
}
\examples{
\dontrun{
import_file_demo_app()
}
}
