% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-file-ext.R
\name{import_delim}
\alias{import_delim}
\title{Wrapper of data.table::fread to import delim files with few presets}
\usage{
import_delim(file, skip, encoding, na.strings)
}
\arguments{
\item{file}{file}

\item{encoding}{encoding}

\item{na.strings}{na.strings}
}
\value{
data.frame
}
\description{
Wrapper of data.table::fread to import delim files with few presets
}
