% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-summary.R
\name{is_consecutive}
\alias{is_consecutive}
\title{Checks if elements in vector are equally spaced as indication of ID}
\usage{
is_consecutive(data)
}
\arguments{
\item{data}{vector}
}
\value{
logical
}
\description{
Checks if elements in vector are equally spaced as indication of ID
}
\examples{
1:10 |> is_consecutive()
sample(1:100,40) |> is_consecutive()
}
