% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_identical_to_previous}
\alias{is_identical_to_previous}
\title{Test if element is identical to the previous}
\usage{
is_identical_to_previous(data, no.name = TRUE)
}
\arguments{
\item{data}{data. vector, data.frame or list}

\item{no.name}{logical to remove names attribute before testing}
}
\value{
logical vector
}
\description{
Test if element is identical to the previous
}
\examples{
c(1, 1, 2, 3, 3, 2, 4, 4) |> is_identical_to_previous()
mtcars[c(1, 1, 2, 3, 3, 2, 4, 4)] |> is_identical_to_previous()
list(1, 1, list(2), "A", "a", "a") |> is_identical_to_previous()
}
