% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_string.R
\name{is_splittable}
\alias{is_splittable}
\title{Determine if any variable in data frame character and contains recognized delimiters}
\usage{
is_splittable(data)
}
\arguments{
\item{data}{vector or data.frame}
}
\value{
logical
}
\description{
Determine if any variable in data frame character and contains recognized delimiters
}
\examples{
any(apply(mtcars, 2, is_splittable))
is_splittable(mtcars)
}
