% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing_page_ui.R
\name{landing_page_ui}
\alias{landing_page_ui}
\title{Tag list of elements for the FreesearchR landing page}
\usage{
landing_page_ui(i18n)
}
\arguments{
\item{i18n}{i18n function loaded in the UI}
}
\value{
tag list
}
\description{
Tag list of elements for the FreesearchR landing page
}
