% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_freesearcheR.R
\name{launch}
\alias{launch}
\title{Easily launch the freesearcheR app}
\usage{
launch(...)
}
\arguments{
\item{...}{passed on to \code{shiny::runApp()}}
}
\value{
shiny app
}
\description{
Easily launch the freesearcheR app
}
