% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_freesearcheR.R
\name{launch_freesearcheR}
\alias{launch_freesearcheR}
\title{Easily launch the freesearcheR app}
\usage{
launch_freesearcheR(...)
}
\arguments{
\item{...}{passed on to \code{shiny::runApp()}}
}
\value{
shiny app
}
\description{
Easily launch the freesearcheR app
}
