% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-file-ext.R
\name{limit_data_size}
\alias{limit_data_size}
\title{Limit the allowed data set size by number of cells}
\usage{
limit_data_size(data, limit = NULL)
}
\arguments{
\item{data}{data.frame}

\item{limit}{cell number limit. Default is NULL.}
}
\value{
data.frame
}
\description{
This function may act to guard a hosted app against very large data sets in
addition to the file size limitations.
The function will limit the data set by dropping rows.
}
\examples{
prod(dim(mtcars))
limit_data_size(mtcars)
limit_data_size(mtcars,100)
}
