% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plot.R
\name{limit_log}
\alias{limit_log}
\title{Easily round log scale limits for nice plots}
\usage{
limit_log(data, fun, ...)
}
\arguments{
\item{data}{data}

\item{fun}{rounding function (floor/ceiling)}

\item{...}{ignored}
}
\value{
numeric vector
}
\description{
Easily round log scale limits for nice plots
}
\examples{
limit_log(-.1, floor)
limit_log(.1, ceiling)
limit_log(-2.1, ceiling)
limit_log(2.1, ceiling)
}
