% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{make_validation}
\alias{make_validation}
\title{Create validation data.frame}
\usage{
make_validation(ls, ...)
}
\arguments{
\item{ls}{validation list}

\item{...}{magic dots}
}
\value{
data.frame
}
\description{
Create validation data.frame
}
\examples{
i18n <- shiny.i18n::Translator$new(translation_csvs_path = here::here("inst/translations"))
i18n$set_translation_language("en")
df_original <- mtcars
df_original[1,2:4] <- NA
df_obs <- df_original |> dplyr::filter(carb==4)
df_vars <- df_original[1:7]
val <- purrr::map2(
  .x = validation_lib(),
  .y = list(
  list(x = df_original, y = df_obs),
  list(x = df_original, y = df_vars),
  list(x=df_original)),
  make_validation
)
val |> make_validation_alerts()

val2 <- purrr::map2(
  .x = validation_lib()[2],
  .y = list(list(x = mtcars, y = mtcars[0])),
  make_validation
)
val2 |> make_validation_alerts()

val3 <- make_validation(
  ls = validation_lib()[[2]],
  list(x = mtcars, y = mtcars[0])
)
}
