% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{make_validation_alerts}
\alias{make_validation_alerts}
\title{Create alert from validation data.frame}
\usage{
make_validation_alerts(data)
}
\arguments{
\item{data}{}
}
\description{
Create alert from validation data.frame
}
