% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{mcar_validate}
\alias{mcar_validate}
\title{MCAR validation based on a gtsummary table bady}
\usage{
mcar_validate(data, outcome = NULL)
}
\arguments{
\item{data}{data}

\item{outcome}{outcome variable}
}
\value{
data.frame
}
\description{
MCAR validation based on a gtsummary table bady
}
