% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{merge_expression}
\alias{merge_expression}
\title{Merge list of expressions}
\usage{
merge_expression(data)
}
\arguments{
\item{data}{list}
}
\value{
expression
}
\description{
Merge list of expressions
}
\examples{
list(
  rlang::call2(.fn = "select", !!!list(c("cyl", "disp")), .ns = "dplyr"),
  rlang::call2(.fn = "default_parsing", .ns = "FreesearchR")
) |> merge_expression()
}
