% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plot.R
\name{merge_long}
\alias{merge_long}
\title{Wrapper to pivot gtsummary table data to long for plotting}
\usage{
merge_long(list, model.names)
}
\arguments{
\item{list}{a custom regression models list}

\item{model.names}{names of models to include}
}
\value{
list
}
\description{
Wrapper to pivot gtsummary table data to long for plotting
}
