% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{missing_fraction}
\alias{missing_fraction}
\title{Get missingsness fraction}
\usage{
missing_fraction(data)
}
\arguments{
\item{data}{data}
}
\value{
numeric vector
}
\description{
Get missingsness fraction
}
\examples{
c(NA, 1:10, rep(NA, 3)) |> missing_fraction()
}
