% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{missings_validate}
\alias{missings_validate}
\title{Validate function of missingness in data}
\usage{
missings_validate(data)
}
\arguments{
\item{data}{data set}
}
\value{
data.frame
}
\description{
Validate function of missingness in data
}
\examples{
df <- mtcars
df[1,2:4] <- NA
missings_validate(df)
}
