% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut-variable-ext.R
\name{names2val}
\alias{names2val}
\title{Set values as names and names as values}
\usage{
names2val(data)
}
\arguments{
\item{data}{data}
}
\value{
named vector
}
\description{
Set values as names and names as values
}
\examples{
names2val(c("Cylinders" = "cyl", "Transmission" = "am", "Gears" = "gear"))
}
