% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{obs_filter_validate}
\alias{obs_filter_validate}
\title{Observations filter test wrapper}
\usage{
obs_filter_validate(before, after)
}
\arguments{
\item{before}{data before}

\item{after}{data after}
}
\value{
vector
}
\description{
Observations filter test wrapper
}
