% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_model.R
\name{outcome_type}
\alias{outcome_type}
\title{Outcome data type assessment}
\usage{
outcome_type(data)
}
\arguments{
\item{data}{data}
}
\value{
outcome type
}
\description{
Outcome data type assessment
}
\examples{
mtcars |>
  default_parsing() |>
  lapply(outcome_type)
}
