% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-summary.R
\name{overview_vars}
\alias{overview_vars}
\title{Create a data overview data.frame ready for sparklines}
\usage{
overview_vars(data)
}
\arguments{
\item{data}{data}
}
\value{
data.frame
}
\description{
Create a data overview data.frame ready for sparklines
}
\examples{
mtcars |> overview_vars()
}
