% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pipe_string}
\alias{pipe_string}
\title{Reduce character vector with the native pipe operator or character string}
\usage{
pipe_string(data, collapse = "|>\\n")
}
\arguments{
\item{data}{list}
}
\value{
character string
}
\description{
Reduce character vector with the native pipe operator or character string
}
\examples{
list(
  "mtcars",
  rlang::call2(.fn = "select", !!!list(c("cyl", "disp")), .ns = "dplyr"),
  rlang::call2(.fn = "default_parsing", .ns = "FreesearchR")
) |>
  lapply(expression_string) |>
  pipe_string() |>
  expression_string("data<-")
}
