% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plot.R
\name{plot.tbl_regression}
\alias{plot.tbl_regression}
\title{Regression coef plot from gtsummary. Slightly modified to pass on arguments}
\usage{
\method{plot}{tbl_regression}(
  x,
  plot_ref = TRUE,
  remove_header_rows = TRUE,
  remove_reference_rows = FALSE,
  ...
)
}
\arguments{
\item{x}{(\code{tbl_regression}, \code{tbl_uvregression})\cr
A 'tbl_regression' or 'tbl_uvregression' object}

\item{plot_ref}{(scalar \code{logical})\cr
plot reference values}

\item{remove_header_rows}{(scalar \code{logical})\cr
logical indicating whether to remove header rows
for categorical variables. Default is \code{TRUE}}

\item{remove_reference_rows}{(scalar \code{logical})\cr
logical indicating whether to remove reference rows
for categorical variables. Default is \code{FALSE}.}

\item{...}{arguments passed to \code{ggstats::ggcoef_plot(...)}}
}
\value{
ggplot object
}
\description{
Regression coef plot from gtsummary. Slightly modified to pass on arguments
}
\examples{
\dontrun{
mod <- lm(mpg ~ ., default_parsing(mtcars))
p <- mod |>
  gtsummary::tbl_regression() |>
  plot(colour = "variable")
}

}
