% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_euler.R
\name{plot_euler_single}
\alias{plot_euler_single}
\title{Easily plot single euler diagrams}
\usage{
plot_euler_single(data, color.palette = "viridis", ...)
}
\value{
ggplot2 object
}
\description{
Easily plot single euler diagrams
}
\examples{
data.frame(
  A = sample(c(TRUE, TRUE, FALSE), 50, TRUE),
  B = sample(c("A", "C"), 50, TRUE),
  C = sample(c(TRUE, FALSE, FALSE, FALSE), 50, TRUE),
  D = sample(c(TRUE, FALSE, FALSE, FALSE), 50, TRUE)
) |> plot_euler_single()
mtcars[c("vs", "am")] |> plot_euler_single("magma")
}
