% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plots.R
\name{plot_sankey_single}
\alias{plot_sankey_single}
\title{Beautiful sankey plot}
\usage{
plot_sankey_single(data, x, y, color.group = "x", colors = NULL)
}
\arguments{
\item{colors}{}
}
\value{
ggplot2 object
}
\description{
Beautiful sankey plot
}
\examples{
ds <- data.frame(g = sample(LETTERS[1:2], 100, TRUE), first = REDCapCAST::as_factor(sample(letters[1:4], 100, TRUE)), last = REDCapCAST::as_factor(sample(letters[1:4], 100, TRUE)))
ds |> plot_sankey_single("first", "last")
ds |> plot_sankey_single("first", "last", color.group = "y")
}
