% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_model.R
\name{possible_functions}
\alias{possible_functions}
\title{Get possible regression models}
\usage{
possible_functions(data, design = c("cross-sectional"))
}
\arguments{
\item{data}{data}
}
\value{
character vector
}
\description{
Get possible regression models
}
\examples{
mtcars |>
  default_parsing() |>
  dplyr::pull("cyl") |>
  possible_functions(design = "cross-sectional")

mtcars |>
  default_parsing() |>
  dplyr::select("cyl") |>
  possible_functions(design = "cross-sectional")
}
