% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{possible_plots}
\alias{possible_plots}
\title{Get possible regression models}
\usage{
possible_plots(data, source_list = supported_plots())
}
\arguments{
\item{data}{data}
}
\value{
character vector
}
\description{
Get possible regression models
}
\examples{
mtcars |>
  default_parsing() |>
  dplyr::pull("cyl") |>
  possible_plots()

mtcars |>
  default_parsing() |>
  dplyr::select("mpg") |>
  possible_plots()
}
