% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_read_shiny_module.R
\name{m_redcap_readUI}
\alias{m_redcap_readUI}
\alias{m_redcap_readServer}
\alias{redcap_demo_app}
\title{Shiny module to browser and export REDCap data}
\usage{
m_redcap_readUI(id, title = TRUE, url = NULL)

m_redcap_readServer(id)

redcap_demo_app()
}
\arguments{
\item{id}{Namespace id}

\item{include_title}{logical to include title}
}
\value{
shiny ui element

shiny server module
}
\description{
Shiny module to browser and export REDCap data

Test app for the redcap_read_shiny_module
}
\examples{
\dontrun{
redcap_demo_app()
}
}
