% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_table.R
\name{regression_table_create}
\alias{regression_table_create}
\title{Create regression summary table}
\usage{
regression_table_create(
  x,
  ...,
  args.list = NULL,
  fun = "gtsummary::tbl_regression",
  theme = c("jama", "lancet", "nejm", "qjecon")
)
}
\arguments{
\item{x}{(list of) regression model}

\item{...}{ignored for now}

\item{args.list}{args.list for the summary function}

\item{fun}{table summary function. Default is "gtsummary::tbl_regression"}

\item{theme}{summary table theme}
}
\value{
gtsummary list object
}
\description{
Create regression summary table
}
