% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_empty_cols}
\alias{remove_empty_cols}
\title{Removes columns with completenes below cutoff}
\usage{
remove_empty_cols(data, cutoff = 0.7)
}
\arguments{
\item{data}{data frame}

\item{cutoff}{numeric}
}
\value{
data frame
}
\description{
Removes columns with completenes below cutoff
}
\examples{
data.frame(a=1:10,b=NA, c=c(2,NA)) |> remove_empty_cols(cutoff=.5)
}
