% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_nested_list}
\alias{remove_nested_list}
\title{Very simple function to remove nested lists, lik ewhen uploading .rds}
\usage{
remove_nested_list(data)
}
\arguments{
\item{data}{data}
}
\value{
data.frame
}
\description{
Very simple function to remove nested lists, lik ewhen uploading .rds
}
\examples{
dplyr::tibble(a = 1:10, b = rep(list("a"), 10)) |> remove_nested_list()
dplyr::tibble(a = 1:10, b = rep(list(c("a", "b")), 10)) |> as.data.frame()
}
