% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_read_shiny_module.R
\name{repeated_instruments}
\alias{repeated_instruments}
\title{Get names of repeated instruments}
\usage{
repeated_instruments(uri, token)
}
\arguments{
\item{uri}{REDCap database uri}

\item{token}{database token}
}
\value{
vector
}
\description{
Get names of repeated instruments
}
