% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plots.R
\name{sankey_ready}
\alias{sankey_ready}
\title{Readying data for sankey plot}
\usage{
sankey_ready(data, x, y, z = NULL, numbers = "count")
}
\arguments{
\item{z}{}
}
\description{
Readying data for sankey plot
}
\examples{
ds <- data.frame(g = sample(LETTERS[1:2], 100, TRUE), first = REDCapCAST::as_factor(sample(letters[1:4], 100, TRUE)), last = sample(c(letters[1:4], NA), 100, TRUE, prob = c(rep(.23, 4), .08)))
ds |> sankey_ready("first", "last")
ds |> sankey_ready("first", "last", numbers = "percentage")
}
