% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{selectPlotVariables}
\alias{selectPlotVariables}
\title{Wrapper for columnSelectInput}
\usage{
selectPlotVariables(data, exclude = NULL, allow_none = TRUE, var_types, ...)
}
\description{
Wrapper for columnSelectInput
}
