% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_string.R
\name{separate_string}
\alias{separate_string}
\title{Separate string wide or long}
\usage{
separate_string(data, col, delim, direction = c("wide", "long"))
}
\arguments{
\item{data}{data}

\item{col}{column}

\item{delim}{delimiter}

\item{direction}{target direction}
}
\value{
data.frame
}
\description{
Separate string wide or long
}
