% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{simple_snake}
\alias{simple_snake}
\title{Simplified version of the snakecase packages to_snake_case}
\usage{
simple_snake(data)
}
\arguments{
\item{data}{character string vector}
}
\value{
vector
}
\description{
Simplified version of the snakecase packages to_snake_case
}
\examples{
c("foo bar", "fooBar21", "!!Foo'B'a-r", "foo_bar", "F  OO bar") |> simple_snake()
}
