% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sort_by}
\alias{sort_by}
\title{Drop-in replacement for the base::sort_by with option to remove NAs}
\usage{
sort_by(x, y, na.rm = FALSE, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{na.rm}{remove NAs}

\item{...}{passed to base_sort_by}
}
\value{
vector
}
\description{
Drop-in replacement for the base::sort_by with option to remove NAs
}
\examples{
sort_by(c("Multivariable", "Univariable"), c("Univariable", "Minimal", "Multivariable"))
}
