% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_string.R
\name{split-string}
\alias{split-string}
\alias{string_split_ui}
\alias{string_split_server}
\alias{modal_string_split}
\title{String split module based on tidyr::separate_}
\usage{
string_split_ui(id)

string_split_server(id, data_r = reactive(NULL))

modal_string_split(
  id,
  title = i18n$t("Split character string"),
  easyClose = TRUE,
  size = "xl",
  footer = NULL
)
}
\arguments{
\item{id}{id}

\item{data_r}{reactive data}

\item{title}{Modal title}

\item{easyClose}{easyClose}

\item{size}{size}

\item{footer}{footer}
}
\value{
A shiny ui module

shiny module server

shiny modal
}
\description{
String split module based on tidyr::separate_
}
