% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_model.R
\name{supported_functions}
\alias{supported_functions}
\title{Implemented functions}
\usage{
supported_functions()
}
\value{
list
}
\description{
Library of supported functions. The list name and "descr" element should be
unique for each element on list.
}
\examples{
supported_functions()
}
