% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plot.R
\name{symmetrical_scale_x_log10}
\alias{symmetrical_scale_x_log10}
\title{Ensure symmetrical plot around 1 on a logarithmic x scale for ratio plots}
\usage{
symmetrical_scale_x_log10(plot, breaks = c(1, 2, 3, 5, 10), ...)
}
\arguments{
\item{plot}{ggplot2 plot}

\item{breaks}{breaks used and mirrored}

\item{...}{ignored}
}
\value{
ggplot2 object
}
\description{
Ensure symmetrical plot around 1 on a logarithmic x scale for ratio plots
}
