% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-summary.R
\name{type_icons}
\alias{type_icons}
\title{Get data type icons}
\usage{
type_icons(x)
}
\arguments{
\item{x}{character vector of data classes}
}
\value{
list
}
\description{
Get data type icons
}
\examples{
"ordinal" |> type_icons()
default_parsing(mtcars) |> sapply(data_type) |> type_icons()
}
