% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_elements.R
\name{ui_elements}
\alias{ui_elements}
\title{FreesearchR UI elements list}
\usage{
ui_elements(selection)
}
\arguments{
\item{selection}{specify element to output}
}
\value{
Shinu UI elements
}
\description{
FreesearchR UI elements list
}
