% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_summary.R
\name{unique_short}
\alias{unique_short}
\title{Create unique short names of character vector items based on index}
\usage{
unique_short(data, max = 15)
}
\arguments{
\item{data}{character vector}

\item{max}{maximum final name length}
}
\value{
character vector
}
\description{
The function will prefer original names, and only append index to long
strings.
}
\examples{
c("kahdleidnsallskdj", "hej") |> unique_short()
}
