% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-variables-ext.R
\name{update-variables}
\alias{update-variables}
\alias{update_variables_ui}
\alias{update_variables_server}
\title{Select, rename and convert variables}
\usage{
update_variables_ui(id, title = "")

update_variables_server(
  id,
  data,
  height = NULL,
  return_data_on_init = FALSE,
  try_silent = FALSE
)
}
\arguments{
\item{id}{Module's ID}

\item{title}{Module's title, if \code{TRUE} use the default title,
use \code{NULL} for no title or a \code{shiny.tag} for a custom one.}

\item{data}{a \code{data.frame} or a \code{reactive} function returning a \code{data.frame}.}

\item{height}{Height for the table.}

\item{return_data_on_init}{Return initial data when module is called.}

\item{try_silent}{logical: should the report of error messages be suppressed?}
}
\value{
A \code{\link[shiny:reactive]{shiny::reactive()}} function returning the updated data.
}
\description{
Select, rename and convert variables
}
