% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-variables-ext.R
\name{update_variables_datagrid}
\alias{update_variables_datagrid}
\title{Modified from the datamods pacakge}
\usage{
update_variables_datagrid(
  data,
  height = NULL,
  selectionId = NULL,
  buttonId = NULL
)
}
\arguments{
\item{data}{data}

\item{height}{height}

\item{selectionId}{selectionId}

\item{buttonId}{buttonId}
}
\description{
Modified from the datamods pacakge
}
\examples{
mtcars |>
  summary_vars() |>
  update_variables_datagrid()

}
