% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validation_lib}
\alias{validation_lib}
\title{Validation library}
\usage{
validation_lib(name = NULL)
}
\arguments{
\item{name}{Index name}
}
\value{
list
}
\description{
Validation library
}
\examples{
validation_lib()
validation_lib("missings")
}
