% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{vars_filter_validate}
\alias{vars_filter_validate}
\title{Variable filter test wrapper}
\usage{
vars_filter_validate(before, after)
}
\arguments{
\item{before}{data before}

\item{after}{data after}
}
\value{
vector
}
\description{
Variable filter test wrapper
}
\examples{
vars_filter_validate(mtcars, mtcars[1:6])
vars_filter_validate(mtcars, mtcars[0])
}
