% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plots.R
\name{vertical_stacked_bars}
\alias{vertical_stacked_bars}
\title{Vertical stacked bar plot wrapper}
\usage{
vertical_stacked_bars(
  data,
  score = "full_score",
  group = "pase_0_q",
  strata = NULL,
  t.size = 10,
  l.color = "black",
  l.size = 0.5,
  draw.lines = TRUE
)
}
\arguments{
\item{t.size}{}
}
\description{
Vertical stacked bar plot wrapper
}
