% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_summary.R
\name{visual-summary}
\alias{visual-summary}
\alias{visual_summary_ui}
\alias{visual_summary_server}
\alias{modal_visual_summary}
\alias{missings_apex_plot}
\alias{visual_summary}
\title{Data correlations evaluation module}
\usage{
visual_summary_ui(id)

visual_summary_server(id, data_r = shiny::reactive(NULL), ...)

modal_visual_summary(
  id,
  title = "Visual overview of data classes and missing observations",
  easyClose = TRUE,
  size = "xl",
  footer = NULL,
  ...
)

missings_apex_plot(data, animation = FALSE, ...)

visual_summary(data, legend.title = NULL, ylab = "Observations", ...)
}
\arguments{
\item{id}{id}

\item{data_r}{reactive data}

\item{...}{optional arguments passed to data_summary_gather()}

\item{title}{title}

\item{easyClose}{easyClose}

\item{size}{modal size}

\item{footer}{modal footer}

\item{data}{data}

\item{legend.title}{Legend title}

\item{ylab}{Y axis label}
}
\value{
Shiny ui module

shiny server

shiny modal

An \code{\link[apexcharter:apexchart]{apexcharter::apexchart()}} \code{htmlwidget} object.

ggplot2 object
}
\description{
Data correlations evaluation module

Visual summary server

Visual summary modal

Plot missings and class with apexcharter. Not in use with FreesearchR.

Ggplot2 data summary visualisation based on visdat::vis_dat.
}
