% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_summary.R
\name{visual_summary}
\alias{visual_summary}
\title{Ggplot2 data summary visualisation based on visdat::vis_dat.}
\usage{
visual_summary(data, legend.title = "Data class", ...)
}
\arguments{
\item{data}{data}

\item{...}{optional arguments passed to data_summary_gather()}
}
\value{
ggplot2 object
}
\description{
Ggplot2 data summary visualisation based on visdat::vis_dat.
}
\examples{
data_demo <- mtcars
data_demo[sample(1:32, 10), "cyl"] <- NA
data_demo[sample(1:32, 8), "vs"] <- NA
visual_summary(data_demo)
visual_summary(data_demo, palette.fun = scales::hue_pal())
visual_summary(dplyr::storms)
visual_summary(dplyr::storms, summary.fun = data_type)
}
