% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{wrap_plot_list}
\alias{wrap_plot_list}
\title{Wrapping}
\usage{
wrap_plot_list(
  data,
  tag_levels = NULL,
  title = NULL,
  axis.font.family = NULL,
  guides = "collect",
  axes = "collect",
  axis_titles = "collect",
  y.axis.percentage = FALSE,
  ...
)
}
\arguments{
\item{data}{list of ggplot2 objects}

\item{tag_levels}{passed to patchwork::plot_annotation if given. Default is NULL}

\item{title}{panel title}

\item{guides}{passed to patchwork::wrap_plots()}

\item{axes}{passed to patchwork::wrap_plots()}

\item{axis_titles}{passed to patchwork::wrap_plots()}

\item{...}{passed to patchwork::wrap_plots()}
}
\value{
list of ggplot2 objects
}
\description{
Wrapping
}
