% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plots.R
\name{align_axes}
\alias{align_axes}
\title{Aligns axes between plots}
\usage{
align_axes(..., x.axis = TRUE, y.axis = TRUE)
}
\arguments{
\item{...}{ggplot2 objects or list of ggplot2 objects}
}
\value{
list of ggplot2 objects
}
\description{
Aligns axes between plots
}
