% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plots.R
\name{all_but}
\alias{all_but}
\title{Select all from vector but}
\usage{
all_but(data, ...)
}
\arguments{
\item{data}{vector}

\item{...}{exclude}
}
\value{
vector
}
\description{
Select all from vector but
}
\examples{
all_but(1:10, c(2, 3), 11, 5)
}
