% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plots.R
\name{clean_common_axis}
\alias{clean_common_axis}
\title{Extract and clean axis ranges}
\usage{
clean_common_axis(p, axis)
}
\arguments{
\item{p}{plot}

\item{axis}{axis. x or y.}
}
\value{
vector
}
\description{
Extract and clean axis ranges
}
