% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{custom_theme}
\alias{custom_theme}
\title{Custom theme based on unity}
\usage{
custom_theme(
  ...,
  version = 5,
  primary = FreesearchR_colors("primary"),
  secondary = FreesearchR_colors("secondary"),
  bootswatch = "united",
  base_font = bslib::font_google("Montserrat"),
  heading_font = bslib::font_google("Public Sans", wght = "700"),
  code_font = bslib::font_google("Open Sans"),
  success = FreesearchR_colors("success"),
  info = FreesearchR_colors("info"),
  warning = FreesearchR_colors("warning"),
  danger = FreesearchR_colors("danger")
)
}
\arguments{
\item{...}{everything passed on to bslib::bs_theme()}
}
\value{
theme list
}
\description{
Custom theme based on unity
}
