% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations-module.R
\name{data-correlations}
\alias{data-correlations}
\alias{data_correlations_ui}
\alias{data_correlations_server}
\title{Data correlations evaluation module}
\usage{
data_correlations_ui(id, ...)

data_correlations_server(id, data, include.class = NULL, cutoff = 0.7, ...)
}
\arguments{
\item{id}{Module id. (Use 'ns("id")')}

\item{...}{arguments passed to toastui::datagrid}

\item{data}{data}

\item{color.main}{main color}

\item{color.sec}{secondary color}
}
\value{
Shiny ui module

shiny server module
}
\description{
Data correlations evaluation module
}
