% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missings-module.R
\name{data-missings}
\alias{data-missings}
\alias{data_missings_ui}
\alias{data_missings_server}
\title{Data correlations evaluation module}
\usage{
data_missings_ui(id)

data_missings_server(id, data, variable, ...)
}
\arguments{
\item{id}{Module id}

\item{data}{data}

\item{output.format}{output format}
}
\value{
Shiny ui module

shiny server module
}
\description{
Data correlations evaluation module
}
