% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_model.R
\name{data_type}
\alias{data_type}
\title{Data type assessment}
\usage{
data_type(data)
}
\arguments{
\item{data}{data}
}
\value{
outcome type
}
\description{
Data type assessment
}
\examples{
mtcars |>
  default_parsing() |>
  lapply(data_type)
c(1, 2) |> data_type()
1 |> data_type()
c(rep(NA, 10)) |> data_type()
sample(1:100, 50) |> data_type()
factor(letters[1:20]) |> data_type()
}
