% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_read_shiny_module.R
\name{is_valid_redcap_url}
\alias{is_valid_redcap_url}
\title{Title}
\usage{
is_valid_redcap_url(url)
}
\arguments{
\item{url}{}
}
\description{
Title
}
\examples{
url <- c(
  "www.example.com",
  "http://example.com",
  "https://redcap.your.inst/api/"
)
is_valid_redcap_url(url)
}
