% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_FreesearchR.R
\name{launch_FreesearchR}
\alias{launch_FreesearchR}
\title{Easily launch the FreesearchR app}
\usage{
launch_FreesearchR(...)
}
\arguments{
\item{...}{passed on to \code{shiny::runApp()}}
}
\value{
shiny app
}
\description{
All data.frames in the global environment will be accessible through the app.
}
\examples{
\dontrun{
data(mtcars)
shiny_FreesearchR(launch.browser = TRUE)
}
}
