% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plots.R
\name{line_break}
\alias{line_break}
\title{Line breaking at given number of characters for nicely plotting labels}
\usage{
line_break(data, lineLength = 20)
}
\arguments{
\item{lineLength}{}
}
\description{
Line breaking at given number of characters for nicely plotting labels
}
