% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_plot.R
\name{plot.tbl_regression}
\alias{plot.tbl_regression}
\title{Regression coef plot from gtsummary. Slightly modified to pass on arguments}
\usage{
\method{plot}{tbl_regression}(x, ...)
}
\arguments{
\item{x}{(\code{tbl_regression}, \code{tbl_uvregression})\cr
A 'tbl_regression' or 'tbl_uvregression' object}

\item{...}{arguments passed to \code{ggstats::ggcoef_plot(...)}}
}
\value{
ggplot object
}
\description{
Regression coef plot from gtsummary. Slightly modified to pass on arguments
}
\examples{
\dontrun{
mod <- lm(mpg ~ ., mtcars)
p <- mod |>
  gtsummary::tbl_regression() |>
  plot(colour = "variable")
}

}
