% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_euler.R
\name{plot_euler}
\alias{plot_euler}
\title{Easily plot euler diagrams}
\usage{
plot_euler(data, pri, sec, ter = NULL, seed = 2103)
}
\arguments{
\item{data}{data}

\item{seed}{seed}

\item{x}{name of main variable}

\item{y}{name of secondary variables}

\item{z}{grouping variable}
}
\value{
patchwork object
}
\description{
Easily plot euler diagrams
}
\examples{
data.frame(
  A = sample(c(TRUE, TRUE, FALSE), 50, TRUE),
  B = sample(c("A", "C"), 50, TRUE),
  C = sample(c(TRUE, FALSE, FALSE, FALSE), 50, TRUE),
  D = sample(c(TRUE, FALSE, FALSE, FALSE), 50, TRUE)
) |> plot_euler("A", c("B", "C"), "D", seed = 4)
mtcars |> plot_euler("vs", "am", seed = 1)
mtcars |> plot_euler("vs", "am", "cyl", seed = 1)
}
