% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plots.R
\name{subset_types}
\alias{subset_types}
\title{Easily subset by data type function}
\usage{
subset_types(data, types, type.fun = outcome_type)
}
\arguments{
\item{data}{data}

\item{types}{desired types}

\item{type.fun}{function to get type. Default is outcome_type}
}
\value{
vector
}
\description{
Easily subset by data type function
}
\examples{
default_parsing(mtcars) |> subset_types("ordinal")
default_parsing(mtcars) |> subset_types(c("dichotomous", "ordinal"))
#' default_parsing(mtcars) |> subset_types("factor",class)
}
