% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plots.R
\name{supported_plots}
\alias{supported_plots}
\title{Implemented functions}
\usage{
supported_plots()
}
\value{
list
}
\description{
Library of supported functions. The list name and "descr" element should be
unique for each element on list.
\itemize{
\item descr: Plot description
\item primary.type: Primary variable data type (continuous, dichotomous or ordinal)
\item secondary.type: Secondary variable data type (continuous, dichotomous or ordinal)
\item secondary.extra: "none" or NULL to have option to choose none.
\item tertiary.type: Tertiary variable data type (continuous, dichotomous or ordinal)
}
}
\examples{
supported_plots() |> str()
}
