% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hbar.R
\name{vertical_stacked_bars}
\alias{vertical_stacked_bars}
\title{Vertical stacked bar plot wrapper}
\usage{
vertical_stacked_bars(
  data,
  score = "full_score",
  group = "pase_0_q",
  strata = NULL,
  t.size = 10,
  l.color = "black",
  l.size = 0.5,
  draw.lines = TRUE
)
}
\arguments{
\item{data}{data.frame}

\item{score}{outcome variable}

\item{group}{grouping variable}

\item{strata}{stratifying variable}

\item{t.size}{text size}
}
\value{
ggplot2 object
}
\description{
Vertical stacked bar plot wrapper
}
