% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plots.R
\name{wrap_plot_list}
\alias{wrap_plot_list}
\title{Wrapping}
\usage{
wrap_plot_list(data, tag_levels = NULL)
}
\arguments{
\item{data}{list of ggplot2 objects}

\item{tag_levels}{passed to patchwork::plot_annotation if given. Default is NULL}
}
\value{
list of ggplot2 objects
}
\description{
Wrapping
}
